+++ src/Driver/DatabaseDriver.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Rougin\Describe\Driver;
+
+use Rougin\Describe\Drivers\DatabaseDriver as BaseDriver;
+
+/**
+ * Database Driver
+ *
+ * A database driver for using available database drivers.
+ * 
+ * @package  Describe
+ * @category Driver
+ * @author   Rougin Royce Gutib <rougingutib@gmail.com>
+ */
+class DatabaseDriver extends BaseDriver {}